+++ lib/api/domain/HttpStatusError.js
@@ -1,8 +1,6 @@
-class HttpStatusError extends Error {
+export default class HttpStatusError extends Error {
   constructor(statusCode) {
     super('Request was not successful');
     this.statusCode = statusCode;
   }
 }
-
-export { HttpStatusError };
