+++ lib/api/domain/Response.js
@@ -1,4 +1,4 @@
-class Response {
+export default class Response {
   constructor({ statusCode, body, headers = {} }) {
     this.statusCode = statusCode;
     this.body = body;
@@ -38,5 +38,3 @@ class Response {
     return new Response({ statusCode: 500 });
   }
 }
-
-export default Response;
