+++ lib/api/index.js
@@ -1,5 +1,5 @@
 import logger from 'lib/infrastructure/logging/logger';
-import { HttpStatusError } from './HttpStatusError';
+import HttpStatusError from './domain/HttpStatusError';
 
 async function request(path, { token, ...options }) {
   const url = `${process.env.NEXT_PUBLIC_API_URL}${path}`;
@@ -21,4 +21,8 @@ async function request(path, { token, ...options }) {
     logger.error(`Fetching ${url} failed`, response.status);
     throw new HttpStatusError(response.status);
   }
+}
+
+export function requestSnapshot(snapshotId, options) {
+  return request(`/snapshots/${snapshotId}`, options);
 }
