+++ lib/api/utils/createEndpoint.js
@@ -1,7 +1,10 @@
-import Response from './response';
+import Response from '../domain/Response';
 import { logger } from 'lib/infrastructure/logging';
 
-export function createEndpoint({ allowedMethods, validators }, endpoint) {
+export default function createEndpoint(
+  { allowedMethods, validators },
+  endpoint
+) {
   return async (req, res) => {
     function send(response) {
       Object.entries(response.headers).forEach(([key, value]) =>
