+++ lib/api/utils/createMockResponse.js
@@ -1,15 +1,15 @@
-export function createMockResponse() {
+export default function createMockResponse() {
   return {
     body: undefined,
     statusCode: undefined,
     headers: {},
-    status: function (statusCode) {
+    status: function(statusCode) {
       this.statusCode = statusCode;
     },
-    setHeader: function (key, value) {
+    setHeader: function(key, value) {
       this.headers[key] = value;
     },
-    send: function (body) {
+    send: function(body) {
       this.body = body;
     }
   };
